requirements.txt
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv~=0.19.0
Flask~=2.1.2